refactor: extract useReminders hook
Moves reminder settings state (reminderSettings, showRemindersSettings,
showMorningTimePicker), the localStorage persist effect, and the
applyReminderPreset / updateCategoryReminder functions out of App.jsx
into a dedicated useReminders hook.
